docs: Modbus map: Add in the modbus map
A document which describes the mapping of Modbus values to data-points
delivered from the heating controller.

A Danfoss MCX06D is used I imagine there's some overlap with whatever the
base Danfoss firmware provides and the additions made by Jifon. Assuming
right now this is an aggregate set of Danfoss + Jifon.

Signed-off-by: Bryan O'Donoghue <pure.logic@nexus-software.ie>